Moved filterbank demo to own class
